Moved CoordinateLicense type to top-level package.
This is being implemented in tandem in the us-street-api client.

Notwithstanding the non-backward compatible nature of this change we
will not be incrementing the major version number since this is, at the
time of this writing, part of an unmerged PR.
